Refactor AuditSeverityApplicationTests
AuditSeverityApplicationTests was not refactored during an early
refactoring of the project, where the name of the project changed from
AuditManagementPortal to AuditSeverity. This prevented Maven from
building the application and hence required a fix.

An additonal GET mapping was also added to the controller to allow for
an easy way to check if the application is running successfully on the
localhost or AWS server.