Add Feign client for AuthorizationService
As the REST endpoint should only be accessed by authorized requests,
there was a need to access the endpoints of the AuthorizationService
in order to validate the request. This will be handled by the new
Feign client. A new class, AuthResponse, was also created to handle
the response from validating the request.

Documentation is also added for some methods.